+++ keanu-project/src/main/java/io/improbable/keanu/algorithms/mcmc/proposal/GaussianProposalDistribution.java
@@ -26,7 +26,7 @@ public class GaussianProposalDistribution implements ProposalDistribution {
     }
 
     @Override
-    public Proposal getProposal(Set<Variable> variables, KeanuRandom random) {
+    public Proposal getProposal(Set<? extends Variable> variables, KeanuRandom random) {
         Proposal proposal = new Proposal();
         for (Variable variable : variables) {
             if (!(variable.getValue() instanceof DoubleTensor)) {