+++ keanu-project/src/main/java/io/improbable/keanu/network/KeanuProbabilisticModel.java
@@ -6,6 +6,7 @@ import io.improbable.keanu.algorithms.ProbabilisticModel;
 import io.improbable.keanu.algorithms.Variable;
 import io.improbable.keanu.algorithms.VariableReference;
 import io.improbable.keanu.algorithms.graphtraversal.VertexValuePropagation;
+import io.improbable.keanu.tensor.dbl.DoubleTensor;
 import io.improbable.keanu.vertices.ProbabilityCalculator;
 import io.improbable.keanu.vertices.Vertex;
 
@@ -14,6 +15,7 @@ import java.util.Collection;
 import java.util.List;
 import java.util.Map;
 import java.util.Set;
+import java.util.stream.Collectors;
 
 import static java.util.stream.Collectors.toMap;
 
@@ -23,11 +25,8 @@ import static java.util.stream.Collectors.toMap;
 public class KeanuProbabilisticModel implements ProbabilisticModel {
 
     private final Map<VariableReference, Vertex> vertexLookup;
-
     private final List<Vertex> latentVertices;
-
     private final List<Vertex> observedVertices;
-
     private final List<Vertex> latentOrObservedVertices;
     private final LambdaSectionSnapshot lambdaSectionSnapshot;
 
@@ -50,7 +49,9 @@ public class KeanuProbabilisticModel implements ProbabilisticModel {
 
     @Override
     public double logProb(Map<VariableReference, ?> inputs) {
-        cascadeValues(inputs);
+        if (!inputs.isEmpty()) {
+            cascadeValues(inputs);
+        }
         return ProbabilityCalculator.calculateLogProbFor(this.latentOrObservedVertices);
     }
 
@@ -72,13 +73,15 @@ public class KeanuProbabilisticModel implements ProbabilisticModel {
 
     @Override
     public double logLikelihood(Map<VariableReference, ?> inputs) {
-        cascadeValues(inputs);
+        if (!inputs.isEmpty()) {
+            cascadeValues(inputs);
+        }
         return ProbabilityCalculator.calculateLogProbFor(this.observedVertices);
     }
 
     @Override
-    public List<? extends Variable> getLatentVariables() {
-        return this.latentVertices;
+    public List<Variable> getLatentVariables() {
+        return (List) this.latentVertices;
     }
 
     public List<Vertex> getLatentVertices() {
@@ -87,6 +90,14 @@ public class KeanuProbabilisticModel implements ProbabilisticModel {
 
     public List<Vertex> getLatentOrObservedVertices() {
         return latentOrObservedVertices;
+    }
+
+    @Override
+    public List<Variable<DoubleTensor, ?>> getContinuousLatentVariables() {
+        return getLatentVariables().stream()
+            .filter(v -> v.getValue() instanceof DoubleTensor)
+            .map(v -> (Variable<DoubleTensor, ?>) v)
+            .collect(Collectors.toList());
     }
 
     private void checkBayesNetInHealthyState() {