+++ keanu-project/src/test/java/io/improbable/keanu/algorithms/mcmc/nuts/NUTSTest.java
@@ -47,7 +47,7 @@ public class NUTSTest {
         ProbabilisticModelWithGradient model = new KeanuProbabilisticModelWithGradient(simpleGaussian);
 
         NUTS nuts = NUTS.builder()
-            .adaptEnabled(false)
+            .adaptStepSizeEnabled(false)
             .initialStepSize(initStepSize)
             .maxTreeHeight(maxTreeHeight)
             .saveStatistics(true)
@@ -80,20 +80,21 @@ public class NUTSTest {
         BayesianNetwork simpleGaussian = MCMCTestDistributions.createSimpleGaussian(mu, sigma, 3);
         ProbabilisticModelWithGradient model = new KeanuProbabilisticModelWithGradient(simpleGaussian);
 
-        int sampleCount = 700;
+        int sampleCount = 300;
         NUTS nuts = NUTS.builder()
-            .adaptCount(sampleCount)
+            .adaptCount(sampleCount / 4)
             .build();
 
         NetworkSamples posteriorSamples = nuts.getPosteriorSamples(
             model,
             model.getLatentVariables(),
             sampleCount
-        ).drop(sampleCount / 4);
+        );
 
         Vertex<DoubleTensor> vertex = simpleGaussian.getContinuousLatentVertices().get(0);
+        List<DoubleTensor> nutsSamples = posteriorSamples.get(vertex).asList();
 
-        MCMCTestDistributions.samplesMatchSimpleGaussian(mu, sigma, posteriorSamples.get(vertex).asList(), 0.1);
+        MCMCTestDistributions.samplesMatchSimpleGaussian(mu, sigma, nutsSamples);
     }
 
     @Test
@@ -183,9 +184,12 @@ public class NUTSTest {
 
     @Category(Slow.class)
     @Test
-    /**
+    /*
      * This test assumes the functional logic of NUTS has not been changed.
      * It simply checks that the samples produced are identical to a previous build.
+     * If this test fails, it may still be functionally correct. Please make sure
+     * you understand why the samples have changed before accepting them by updating these
+     * magic numbers.
      */
     public void checksSamplesAgainstMagicNumbers() {
         double mu = 0.0;
@@ -196,6 +200,7 @@ public class NUTSTest {
         NUTS nuts = NUTS.builder()
             .adaptCount(5)
             .targetAcceptanceProb(0.65)
+            .maxTreeHeight(10)
             .build();
 
         NetworkSamples posteriorSamples = nuts.getPosteriorSamples(
@@ -215,25 +220,57 @@ public class NUTSTest {
         //    System.out.println("Assert.assertEquals(" + samples.get(i).scalar() + ", samples.get(" + i + ").scalar(), epsilon);");
         //}
 
-        Assert.assertEquals(3.0, samples.get(0).scalar(), epsilon);
-        Assert.assertEquals(3.0, samples.get(1).scalar(), epsilon);
-        Assert.assertEquals(-3.340614624594811, samples.get(2).scalar(), epsilon);
-        Assert.assertEquals(-0.6395789896740856, samples.get(3).scalar(), epsilon);
-        Assert.assertEquals(-1.1018237282635894, samples.get(4).scalar(), epsilon);
-        Assert.assertEquals(-1.1018237282635894, samples.get(5).scalar(), epsilon);
-        Assert.assertEquals(-1.1018237282635894, samples.get(6).scalar(), epsilon);
-        Assert.assertEquals(-0.29068491812065367, samples.get(7).scalar(), epsilon);
-        Assert.assertEquals(0.821679170528079, samples.get(8).scalar(), epsilon);
-        Assert.assertEquals(-0.597238866015397, samples.get(9).scalar(), epsilon);
-        Assert.assertEquals(1.4029546134497397, samples.get(10).scalar(), epsilon);
-        Assert.assertEquals(1.4937799961213152, samples.get(11).scalar(), epsilon);
-        Assert.assertEquals(-1.5598695524469481, samples.get(12).scalar(), epsilon);
-        Assert.assertEquals(-0.5122490976946052, samples.get(13).scalar(), epsilon);
-        Assert.assertEquals(-3.087992058616404, samples.get(14).scalar(), epsilon);
-        Assert.assertEquals(-0.5210686093406434, samples.get(15).scalar(), epsilon);
-        Assert.assertEquals(-0.7600466031184221, samples.get(16).scalar(), epsilon);
-        Assert.assertEquals(0.37494461240313726, samples.get(17).scalar(), epsilon);
-        Assert.assertEquals(0.022250923217856877, samples.get(18).scalar(), epsilon);
-        Assert.assertEquals(0.022250923217856877, samples.get(19).scalar(), epsilon);
+        Assert.assertEquals(2.655769904078054, samples.get(0).scalar(), epsilon);
+        Assert.assertEquals(2.655769904078054, samples.get(1).scalar(), epsilon);
+        Assert.assertEquals(-1.0165440910526553, samples.get(2).scalar(), epsilon);
+        Assert.assertEquals(-1.0165440910526553, samples.get(3).scalar(), epsilon);
+        Assert.assertEquals(-1.1033732018241589, samples.get(4).scalar(), epsilon);
+        Assert.assertEquals(-1.5926775601489656, samples.get(5).scalar(), epsilon);
+        Assert.assertEquals(-1.5926775601489656, samples.get(6).scalar(), epsilon);
+        Assert.assertEquals(0.06269495259568081, samples.get(7).scalar(), epsilon);
+        Assert.assertEquals(0.06269495259568081, samples.get(8).scalar(), epsilon);
+        Assert.assertEquals(0.06269495259568081, samples.get(9).scalar(), epsilon);
+        Assert.assertEquals(-1.9067350000491319, samples.get(10).scalar(), epsilon);
+        Assert.assertEquals(-0.16217717982790925, samples.get(11).scalar(), epsilon);
+        Assert.assertEquals(-0.16217717982790925, samples.get(12).scalar(), epsilon);
+        Assert.assertEquals(-1.7749397804101408, samples.get(13).scalar(), epsilon);
+        Assert.assertEquals(-1.95774120776884, samples.get(14).scalar(), epsilon);
+        Assert.assertEquals(1.1167543688694772, samples.get(15).scalar(), epsilon);
+        Assert.assertEquals(-2.3326358434845753, samples.get(16).scalar(), epsilon);
+        Assert.assertEquals(-2.425061791760318, samples.get(17).scalar(), epsilon);
+        Assert.assertEquals(1.3182493371930095, samples.get(18).scalar(), epsilon);
+        Assert.assertEquals(-1.3168733581639154, samples.get(19).scalar(), epsilon);
+
     }
+
+    @Test(expected = IllegalArgumentException.class)
+    public void doesValidateAdapStepSizeCount() {
+        NUTS.builder().adaptCount(-1);
+    }
+
+    @Test(expected = IllegalArgumentException.class)
+    public void doesValidateInitialStepSize() {
+        NUTS.builder().initialStepSize(-0.1);
+    }
+
+    @Test(expected = IllegalArgumentException.class)
+    public void doesValidateTargetAcceptanceUpper() {
+        NUTS.builder().targetAcceptanceProb(1.1);
+    }
+
+    @Test(expected = IllegalArgumentException.class)
+    public void doesValidateTargetAcceptanceLower() {
+        NUTS.builder().targetAcceptanceProb(-0.1);
+    }
+
+    @Test(expected = IllegalArgumentException.class)
+    public void doesValidateMaxEnergyChange() {
+        NUTS.builder().maxEnergyChange(-0.1);
+    }
+
+    @Test(expected = IllegalArgumentException.class)
+    public void doesValidateMaxTreeHeight() {
+        NUTS.builder().maxTreeHeight(0);
+    }
+
 }